Add files via upload
Code seems stable.  Need to do some cleanup with the display and code in general.  Project is taking a slight turn.  These CD players do not list the track on the main display.  They expect you to have a monitor hooked up to the HDMI port. Therefore, this project may be turning into more of a remote control / remote display.  Other issues include not being able to track up/down without it beginning to immediately play.  If we are going to have to make this a one ESP per CD player, then I may just use serial control instead of Ethernet network.  Will most likely be cheaper but open to suggestions.

kw